Added Save/Load feature as well as other needed button functions.
Did some house cleaning as well in hopes to make code more clear. Still
need to add a button to handle how the user gets their character/stat
points. This will involve using one of the three provided methods in the
rulebook for Cyberpunk 2020. Also need to save skills to the .char file.