Refactored the Ammunition hierarchy.
Added a Type class to denote the type of ammunition as well as a Name
class for general naming purposes. Can be used to name items as well as
parts of an Ammunition.

Thinking that maybe Type and Name can be merged since Type is a category
name; the behavior between the two is also identical.