Created the base tab which the biography info is displayed.
Also, implemented the first half of the tab that represents the basic
statistics, name, and age

This commit marks the start of following the Coding Style present in the
wiki on GitHub. This also means updating the previous code ONLY when it
MUST be modified. It will otherwise remain the same as the old Coding
Style.